requirements.txt
numpy
scipy
open3d
Pillow
imageio
pyntcloud
jupyterlab
matplotlib
scikit-image
opencv-python
torch==1.8.0+cu111
torchvision==0.9.0+cu111
torchaudio==0.8.0
